T85378: When decoding values, check the supplied value type
The Cassandra Node.JS drivers seems to be caching values when retrieving
rows from the DB, so sometimes integers and decimals are returned by the
driver as JS Number instances, and not Buffer ones. This commit
introduces a check in both decodeVarInt() and decodeDecimal() to ensure
the right value type is provided.